# improve ByteSize function