old:
```
Benchmark_ByteSize/0-12  	394620337	         2.983 ns/op	       0 B/op	       0 allocs/op
Benchmark_ByteSize/0-12  	402809990	         2.958 ns/op	       0 B/op	       0 allocs/op
Benchmark_ByteSize/0-12  	413246502	         2.921 ns/op	       0 B/op	       0 allocs/op
Benchmark_ByteSize/0-12  	403580769	         2.928 ns/op	       0 B/op	       0 allocs/op
Benchmark_ByteSize/1-12  	12665492	        99.71 ns/op	      26 B/op	       2 allocs/op
Benchmark_ByteSize/1-12  	12164020	        98.69 ns/op	      26 B/op	       2 allocs/op
Benchmark_ByteSize/1-12  	12065070	       101.4 ns/op	      26 B/op	       2 allocs/op
Benchmark_ByteSize/1-12  	12394597	        98.16 ns/op	      26 B/op	       2 allocs/op
Benchmark_ByteSize/500-12         	12344122	        97.39 ns/op	      28 B/op	       2 allocs/op
Benchmark_ByteSize/500-12         	12245131	        97.55 ns/op	      28 B/op	       2 allocs/op
Benchmark_ByteSize/500-12         	12014792	        97.59 ns/op	      28 B/op	       2 allocs/op
Benchmark_ByteSize/500-12         	12355567	        97.73 ns/op	      28 B/op	       2 allocs/op
Benchmark_ByteSize/1024-12        	12244044	        98.73 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1024-12        	12220083	       101.0 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1024-12        	11133416	       102.3 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1024-12        	12035841	       101.2 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1126-12        	10641859	       111.8 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1126-12        	10875364	       111.6 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1126-12        	10740609	       120.4 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1126-12        	10505193	       113.0 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1048576-12     	11958792	       100.0 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1048576-12     	11678012	       102.1 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1048576-12     	11683394	       101.4 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1048576-12     	11338077	        98.68 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1153024-12     	10879534	       112.3 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1153024-12     	10741414	       112.3 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1153024-12     	10648194	       111.2 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1153024-12     	10915987	       111.4 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1073741824-12  	12326180	        99.14 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1073741824-12  	11839748	       102.8 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1073741824-12  	11915380	       101.2 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1073741824-12  	11987697	        97.93 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1180696576-12  	10691478	       111.6 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1180696576-12  	10902478	       111.4 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1180696576-12  	10897396	       111.6 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1180696576-12  	10656692	       111.8 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1099511627776-12         	12218870	       100.1 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1099511627776-12         	11694182	       101.6 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1099511627776-12         	11744306	       100.5 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1099511627776-12         	12106914	       106.1 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1209033293824-12         	10924442	       111.5 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1209033293824-12         	10612323	       112.8 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1209033293824-12         	10885879	       113.1 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1209033293824-12         	10919641	       120.9 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1125899906842624-12      	12151698	       102.2 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1125899906842624-12      	11863318	       100.6 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1125899906842624-12      	11944686	        99.75 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1125899906842624-12      	11987617	        97.80 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1238050092875776-12      	10737126	       111.2 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1238050092875776-12      	11016943	       111.6 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1238050092875776-12      	10792494	       115.3 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1238050092875776-12      	10666341	       113.5 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1152921504606846976-12   	11189354	       102.0 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1152921504606846976-12   	11548677	       101.7 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1152921504606846976-12   	12055868	       106.0 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1152921504606846976-12   	11992898	        98.86 ns/op	      27 B/op	       2 allocs/op
Benchmark_ByteSize/1267763295104794624-12   	10693737	       110.8 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1267763295104794624-12   	10881495	       111.0 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1267763295104794624-12   	10967534	       116.0 ns/op	      29 B/op	       2 allocs/op
Benchmark_ByteSize/1267763295104794624-12   	10403409	       115.0 ns/op	      29 B/op	       2 allocs/op
```
new:
```
Benchmark_ByteSize/0-12  	381202279	         2.953 ns/op	       0 B/op	       0 allocs/op
Benchmark_ByteSize/0-12  	412914474	         2.923 ns/op	       0 B/op	       0 allocs/op
Benchmark_ByteSize/0-12  	392389334	         2.992 ns/op	       0 B/op	       0 allocs/op
Benchmark_ByteSize/0-12  	411269878	         2.926 ns/op	       0 B/op	       0 allocs/op
Benchmark_ByteSize/1-12  	66645376	        18.02 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1-12  	67031926	        18.48 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1-12  	65031100	        18.79 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1-12  	67271643	        18.37 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/500-12         	55519891	        21.61 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/500-12         	56071926	        21.70 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/500-12         	52847816	        21.71 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/500-12         	56364048	        21.67 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1024-12        	60792967	        19.33 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1024-12        	61283637	        19.65 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1024-12        	59079831	        20.01 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1024-12        	58885951	        19.84 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1126-12        	52129024	        22.86 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1126-12        	50120601	        22.93 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1126-12        	51379209	        22.75 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1126-12        	51349436	        22.64 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1048576-12     	62054118	        19.15 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1048576-12     	61777123	        19.36 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1048576-12     	63581073	        19.94 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1048576-12     	63366613	        19.45 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1153024-12     	52423779	        22.41 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1153024-12     	52352594	        22.35 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1153024-12     	48861012	        22.79 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1153024-12     	41920482	        24.33 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1073741824
Benchmark_ByteSize/1073741824-12  	61493130	        19.40 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1073741824-12  	59481151	        19.84 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1073741824-12  	61494837	        19.85 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1073741824-12  	60446295	        19.83 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1180696576-12  	52824068	        22.09 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1180696576-12  	52169942	        22.17 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1180696576-12  	55176746	        22.52 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1180696576-12  	54001458	        22.39 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1099511627776-12         	61136632	        18.58 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1099511627776-12         	61396255	        18.96 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1099511627776-12         	63482826	        19.51 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1099511627776-12         	62977662	        19.30 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1209033293824-12         	53747990	        22.98 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1209033293824-12         	52719302	        22.03 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1209033293824-12         	53753505	        21.85 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1209033293824-12         	54607508	        22.76 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1125899906842624-12      	65190524	        18.50 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1125899906842624-12      	64880208	        19.68 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1125899906842624-12      	64009957	        20.05 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1125899906842624-12      	66231104	        18.61 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1238050092875776-12      	55698992	        21.87 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1238050092875776-12      	54699502	        22.81 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1238050092875776-12      	55585365	        21.98 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1238050092875776-12      	55902247	        21.81 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1152921504606846976-12   	67190190	        18.43 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1152921504606846976-12   	66749640	        19.93 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1152921504606846976-12   	65533930	        18.57 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1152921504606846976-12   	62504475	        18.38 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1267763295104794624-12   	54216764	        21.37 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1267763295104794624-12   	55445602	        21.37 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1267763295104794624-12   	52494489	        21.50 ns/op	      16 B/op	       1 allocs/op
Benchmark_ByteSize/1267763295104794624-12   	56084814	        21.37 ns/op	      16 B/op	       1 allocs/op
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **Refactor**
  * Improved byte size formatting using integer arithmetic for more accurate and efficient output, with safeguards for very large values.

* **Tests**
  * Added benchmark tests to measure performance and memory usage of byte size formatting.
  * Introduced tests validating behavior near maximum supported input values.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->